Fix numpy serialization in quant APIs + stress test button UX
- Wrap round() with float() in all to_dict() methods to prevent
  numpy.float64 from reaching FastAPI jsonable_encoder (500 error)
- Fix stress test button: show error message instead of disappearing
- Affected: var.py, garch.py, portfolio_risk.py, portfolio_optimizer.py,
  quant_signals.py, stress_test.py

+++ backend/app/risk/garch.py
@@ -26,13 +26,13 @@ class GARCHResult:
 
     def to_dict(self) -> dict:
         return {
-            "current_vol": round(self.current_vol, 6),
-            "forecast_1": round(self.forecast_1, 6),
-            "forecast_5": round(self.forecast_5, 6),
-            "long_run_vol": round(self.long_run_vol, 6),
+            "current_vol": float(round(self.current_vol, 6)),
+            "forecast_1": float(round(self.forecast_1, 6)),
+            "forecast_5": float(round(self.forecast_5, 6)),
+            "long_run_vol": float(round(self.long_run_vol, 6)),
             "method": self.method,
-            "alpha": round(self.alpha, 4),
-            "beta": round(self.beta, 4),
+            "alpha": float(round(self.alpha, 4)),
+            "beta": float(round(self.beta, 4)),
         }
 
     @property
